removing absolute paths and moving tinyxml into separate dir
