+++ tinyxml/tinystr.cpp
