+++ tinyxml/tinystr.h
