+++ tinyxml/tinyxmlerror.cpp
